requirements.txt
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.4.0
streamlit==1.12.0
